# swagger.yaml
swagger: "2.0"
info:
  description: "API documentation for the Sunglasses Store application"
  version: "1.0.0"
  title: "Sunglasses Store API"
host: "localhost:3001"
basePath: "/"
schemes:
  - "http"
paths:
...
  /users/register:
    post:
      summary: "Register a new user"
      parameters:
        - name: "user"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              username:
                type: "string"
              password:
                type: "string"
      responses:
        201:
          description: "User registered successfully"
  /users/login:
    post:
      summary: "Login a user"
      parameters:
        - name: "user"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              username:
                type: "string"
              password:
                type: "string"
      responses:
        200:
          description: "User logged in successfully"
          schema:
            type: "object"
            properties:
              accessToken:
                type: "string"
        403:
          description: "Invalid username or password"
...
securityDefinitions:
  BearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"